split MINIT for Locale
